Extract TextFormatBase as superclass for HTTPTextFormat and RackHTTPTextFormat
